cleanup and use built in confirm

+++ priv/www/js/projects/projects.js
@@ -1,33 +1,10 @@
-$(function () {
-  var projectName = $("#project-name"),
-    projectDescription = $("#project-description"),
-    tips = $(".validate-tips"),
-    allFields = $([]).add(projectName).add(projectDescription);
+function addProjectDialog() {
+  var projectName = $("#project-name");
+  var projectDescription = $("#project-description");
+  var tips = $(".validate-tips");
+  var allFields = $([]).add(projectName).add(projectDescription);
   
-  function populateProjectsTable() {
-    // I use a bogus query string to ensure that the browser
-    // won't replace the html version with the json verson
-    // from this resource. There is probably a better way.
-    var url = "/projects/index";
-    
-    $.get(url, function(projects) {
-      $('tbody').empty();
-      $('tbody').html(projects);
-      $('.configure-button').button({
-        icons: {primary: "ui-icon-wrench"}
-      });
-      $('.delete-button').button({
-        icons: {primary: "ui-icon-trash"}
-      }).click(function() {
-        toDelete = $(this).attr("id");
-        $('#delete-dialog').dialog("open");
-      });
-    });
-  }
-  
-  populateProjectsTable();
-  
-  $("#add-dialog").dialog({
+  var dialog = $("#add-dialog").dialog({
     autoOpen: false,
     modal: true,
     buttons: {
@@ -64,36 +41,53 @@ $(function () {
     }
   });
   
-  $("#delete-dialog").dialog({
-    autoOpen: false,
-    modal: true,
-    buttons: {
-      "Delete project": function() {
-        $.ajax({
-          type: "DELETE", 
-          url: "/projects/" + toDelete,
-          dataType: "json",
-          contentType: "application/json",
-          complete: function(req, status) {
-            if (req.status == 204) {
-              populateProjectsTable();
-            } else {
-              alert("An error occurred" + req.status);
-            }
-          }
-        });
-        $(this).dialog("close");
-      },
-      Cancel: function() {
-        $(this).dialog("close");
+  return dialog;
+}
+
+function deleteProject(id) {
+  if (confirm("Are you sure? This is permanent.")) {
+    $.ajax({
+      type: "DELETE", 
+      url: "/projects/" + id,
+      dataType: "json",
+      contentType: "application/json",
+      complete: function(req, status) {
+        if (req.status == 204) {
+          populateProjectsTable();
+        } else {
+          alert("An error occurred" + req.status);
+        }
       }
-    }
+    });
+  }
+}
+
+function populateProjectsTable() {
+  var url = "/projects/index";
+  
+  $.get(url, function(projects) {
+    $('tbody').empty();
+    $('tbody').html(projects);
+    $('.configure-button').button({
+      icons: {primary: "ui-icon-wrench"}
+    });
+    $('.delete-button').button({
+      icons: {primary: "ui-icon-trash"}
+    }).click(function(e) {
+      id = $(e.target).attr("id");
+      deleteProject(id);
+      $('#delete-dialog').dialog("open");
+    });
   });
+}
+
+$(function () {
+  populateProjectsTable();
   
   $("#create-project").button({
     icons: {primary: "ui-icon-plus"}
   }).click(function() {
-    $("#add-dialog").dialog("open");
+    addProjectDialog().dialog("open");
   });
   
 });
